Move the blueprint manager into the top level package
