+++ src/brad/admin/bootstrap_schema.py
@@ -6,7 +6,7 @@ from brad.blueprint.sql_gen.table import TableSqlGenerator
 from brad.config.engine import Engine
 from brad.config.file import ConfigFile
 from brad.planner.data import bootstrap_blueprint
-from brad.front_end.blueprint_manager import BlueprintManager
+from brad.blueprint_manager import BlueprintManager
 from brad.front_end.engine_connections import EngineConnections
 
 logger = logging.getLogger(__name__)
