+++ src/brad/admin/modify_blueprint.py
@@ -9,7 +9,7 @@ from brad.blueprint.sql_gen.table import (
 from brad.config.engine import Engine
 from brad.config.file import ConfigFile
 from brad.planner.enumeration.blueprint import EnumeratedBlueprint
-from brad.front_end.blueprint_manager import BlueprintManager
+from brad.blueprint_manager import BlueprintManager
 from brad.front_end.engine_connections import EngineConnections
 
 logger = logging.getLogger(__name__)
