+++ src/brad/routing/location_aware_round_robin.py
@@ -1,5 +1,5 @@
 from brad.config.engine import Engine
-from brad.front_end.blueprint_manager import BlueprintManager
+from brad.blueprint_manager import BlueprintManager
 from brad.routing.router import Router
 from brad.query_rep import QueryRep
