+++ src/brad/routing/rule_based.py
@@ -7,7 +7,7 @@ from importlib.resources import files, as_file
 import brad.routing
 from brad.blueprint import Blueprint
 from brad.config.engine import Engine
-from brad.front_end.blueprint_manager import BlueprintManager
+from brad.blueprint_manager import BlueprintManager
 from brad.daemon.monitor import Monitor
 from brad.routing.router import Router
 from brad.query_rep import QueryRep
